[yagin][encapsulation for table version 2]
