+++ apps/docs/src/pages/components/table-v2.mdx
@@ -1,0 +1,137 @@
+import {Callout} from 'nextra-theme-docs'
+import styles from '../../styles/Home.module.css'
+import {faThumbsUp, faVirus} from '@fortawesome/free-solid-svg-icons'
+import Textarea from '@comfortdelgro/react-compass/textarea'
+
+# Table-v2
+
+Table version 2 is a wrapper around the core table logic. Most of its job is related to managing state the "react" way, providing types and the rendering implementation of cell/header/footer templates.
+
+---
+
+### Import:
+
+```js
+import {ReactTable} from '@comfortdelgro/react-compass'
+```
+
+or
+
+```js
+import ReactTable from '@comfortdelgro/react-compass/textarea'
+```
+
+---
+
+### Basic table-v2:
+
+<section className={styles.textAreaShowcaseContainer}>
+  <Textarea placeholder='Enter your message' />
+  <Textarea label='Text Label' placeholder='Enter your message' isDisabled/>
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    value='Lorem ipsum dolor sit amet.'
+
+/>
+
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    value='Lorem ipsum dolor sit amet.'
+  />
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    maxLength={50}
+    value={'Lorem ipsum dolor sit amet.'}
+    onChange={() => {}}
+  />
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    maxLength={50}
+    errored
+  />
+</section>
+
+```js
+  <Textarea placeholder='Enter your message' />
+  <Textarea label='Text Label' placeholder='Enter your message' isDisabled/>
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    value='Lorem ipsum dolor sit amet.'
+  />
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    maxLength={50}
+    value={'Lorem ipsum dolor sit amet.'}
+    onChange={() => {}}
+  />
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    maxLength={50}
+    errored
+  />
+```
+
+---
+
+### Custom styling:
+
+You can style React Compass's components with the css prop. <br/>
+With css prop, you can basically do anything that css can do.
+
+<section className={styles.avatarShowcaseContainer}>
+  <Textarea
+    label='Text Label'
+    placeholder='Enter your message'
+    wordCount
+    maxLength={50}
+    errored
+    css={{height: '10em', width: '50em'}}
+  />
+</section>
+
+```js
+<Textarea
+  label='Text Label'
+  placeholder='Enter your message'
+  wordCount
+  maxLength={50}
+  errored
+  css={{height: '10em', width: '50em'}}
+/>
+```
+
+---
+
+### Props:
+
+| Name          | Type               | Default | Description                              |
+| :------------ | :----------------- | :------ | :--------------------------------------- |
+| `placeholder` | `string` \| `null` | `null`  | `Placeholder of the text field`          |
+| `label`       | `string` \| `null` | `null`  | `Label of the text field`                |
+| `errored`     | `true` \| `false`  | `false` | `Throw error when input value is wrong`  |
+| `value`       | `string`           | —       | `Input value`                            |
+| `isDisabled`  | `false`\| `true`   | `false` | `Disable the text field when true`       |
+| `css `        | `CSS`              | —       |                                          |
+| `wordCount`   | `false`\| `true`   | `false` | `Show word count of the input when true` |
+| `maxLength`   | `number`           | —       | `Limit length of the input`              |
+| `onChange`    | `function`         | —       |                                          |
+
+---
+
+<Callout>
+  Check out our [ladle](/ladle?story=textarea--variants) for more info on this
+  component.
+</Callout>
