+++ packages/react-compass/src/index.ts
@@ -194,8 +194,9 @@ export type {
 } from './subBanner'
 export {default as Table} from './table'
 export type {TableProps} from './table'
-export {default as TableV2} from './table-v2'
+export {default as ReactTable} from './table-v2'
 export type {ReactTableProps} from './table-v2'
+export {ReactTableConfig} from './table-v2/react-table-config-model'
 export {default as Tabs} from './tabs'
 export type {TabsProps} from './tabs'
 export {default as TagBox} from './tag-box'
