--- packages/react-compass/src/table-v2/react-table-config-model.ts
@@ -1,5 +1,0 @@
-export interface TableConfig {
-  hasFilter: boolean
-  hasSorted: boolean
-  hasMultipleSorted: boolean
-}
